git refactor HomeScreen
